[PP-01] Add some API about category
