--- frontend/src/App.js
@@ -1,17 +1,0 @@
-import logo from './logo.svg';
-import './App.css';
-
-function App() {
-  return (
-    <div className="App">
-      <header className="App-header">
-        <img src={logo} className="App-logo" alt="logo" />
-        <p>
-            Hello world
-        </p>
-      </header>
-    </div>
-  );
-}
-
-export default App;
